[#step3] chore: 한 클래스당 한 파일로 분리 및 클래스이름에 Plus 삭제
	이유: 하나의 파일당 하나의 클래스만 있어야만 분별하기 쉽기때문에 파일을 분리했고,
	      Plus 라는 이름이 굳이 클래스에 없어도 이해가 되는 것 같아 Plus를 삭제함
